feat(esm): moving event bridge to csm
