+++ infrastructure/parser-graphql-wrapper/src/dynamodb.ts
@@ -1,5 +1,5 @@
 import { Construct } from 'constructs';
-import { config } from './config';
+import { config } from './config/index.ts';
 import {
   ApplicationDynamoDBTable,
   ApplicationDynamoDBTableCapacityMode,
