+++ infrastructure/pocket-event-bridge/package.json
@@ -5,6 +5,7 @@
   "license": "MPL-2.0",
   "main": "dist/main.js",
   "types": "dist/main.ts",
+  "type": "module",
   "scripts": {
     "build": "rm -rf dist && tsc",
     "compile": "tsc --pretty",
