+++ infrastructure/pocket-event-bridge/src/event-rules/account-delete-monitor/config.ts
@@ -1,4 +1,4 @@
-import { config as globalConfig } from '../../config';
+import { config as globalConfig } from '../../config/index.ts';
 
 export const config = {
   queueCheckDelete: {
